requirements.txt
torch
flask
gunicorn
python-dotenv
pandas
transformers
sentence-transformers
